First Cut of console solution
This commit lays out the structure of the projects in the solution which
consists of a main solution project and a test project that users Nunit.
It describes the Interfaces, Classes and Tests to be constructed.  Sets
up DI to enable testing and attempts to separate code into small easily
managed chunks.