Second Cut of console solution
This commit includes two methods for calculating the occurrence of the
words in the text.  At this stage there is a large difference between
regex and split proposed solution but I guess this very much depends on
the definition of a 'word' in this context.  More work required on this.
Two unit tests have been added.