Final Cut of Console Code
Updated tests and added some try catch statements.

I have not removed any of the redundant code that was cut when trying
out the various ideas, this is deliberate so you can see the thought
processes and code tests behind each.  I've left the program.cs so that
it utilising the most efficient path out of those investigated.